refactored added metrics with BaseLoglikelihoodMetric
