+++ src/eval_framework/metrics/loglikelihood/dcs.py
@@ -1,17 +1,12 @@
-import numpy as np
-
-from eval_framework.metrics.base import BaseMetric, MetricResult
-from eval_framework.shared.types import Loglikelihood
+from eval_framework.metrics.base import MetricResult
+from eval_framework.metrics.loglikelihood.base import BaseLoglikelihoodMetric
 
 
-class DistributionalCorrectnessScore(BaseMetric[Loglikelihood]):
+class DistributionalCorrectnessScore(BaseLoglikelihoodMetric):
     NAME = "Distributional Correctness Score"
 
     LC: float = 1.0 # Default reward weight for correct answers
     LW: float = 1.0 # Default penalty weight for wrong answers
-
-    def _normalise_text(self, text: str) -> str:
-        return text.strip().lower()
 
     def __init__(
         self,
@@ -20,27 +15,13 @@ class DistributionalCorrectnessScore(BaseMetric[Loglikelihood]):
         lw: float | None = None,
         assume_len_normalised: bool = False,
     ) -> None:
+        super().__init__(assume_len_normalised=assume_len_normalised)
         self._lc = float(lc) if lc is not None else float(self.LC)
         self._lw = float(lw) if lw is not None else float(self.LW)
         if not (self._lc >= 0 and self._lw >= 0 and self._lc >= self._lw):
             raise ValueError(
                 f"Invalid DCS loadings: lc={self._lc}, lw={self._lw}. Require lc>=0, lw>=0, and lc>=lw."
             )
-        self._assume_len_normalised = assume_len_normalised
-
-    def _length_normalise_loglikelihoods(self, loglikelihoods: dict) -> dict:
-        output = {}
-        for k, v in loglikelihoods.items():
-            length = len(k)
-            output[k] = v / length if length > 0 else v
-        return output
-
-    def _softmax(self, log_probs: dict) -> dict:
-        vals = list(log_probs.values())
-        m = max(vals)
-        exp_vals = [math.exp(x - m) for x in vals]
-        total = sum(exp_vals)
-        return {k: ev / total for k, ev in zip(log_probs.keys(), exp_vals)}
 
     def calculate(self, response: Loglikelihood) -> list[MetricResult]:
         if response.error is not None:
